refactor(mux): isolate reconcile mount proof
Keep structural planning pure and centralize native pane role precedence.
